requirements.txt
streamlit
pandas
numpy
scikit-learn
scikit-surprise
plotly
fuzzywuzzy
python-Levenshtein
openpyxl